+++ packages/terra-button/src/terra-dev-site/doc/example/ButtonBlock.jsx
@@ -5,7 +5,7 @@ import Button from 'terra-button/lib/Button';
 const buttonStyle = { margin: '5px' };
 
 const ButtonBlock = () => (
-  <Button text="Default Block" isBlock style={buttonStyle} />
+  <Button text="Neutral Block" isBlock style={buttonStyle} />
 );
 
 export default ButtonBlock;