+++ packages/terra-button/src/terra-dev-site/test/button/Variants/GhostButton.test.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import Button from '../../../../Button';
+
+const GhostButton = () => (
+  <div id="ghost" style={{ padding: '5px' }}>
+    <Button id="ghostButton" text="Ghost" variant="ghost" />
+    <Button id="ghostButtonDisabled" text="Ghost Disabled" variant="ghost" isDisabled />
+  </div>
+);
+
+export default GhostButton;