+++ packages/terra-action-footer/CHANGELOG.md
@@ -4,6 +4,12 @@ ChangeLog
 Unreleased
 ----------
 
+2.10.0 - (May 1, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+* Updated WDIO tests to use themeCombinationOfCustomProperties instead of themeEachCustomProperty
+
 2.9.0 - (April 16, 2019)
 ------------------
 ### Changed