+++ packages/terra-search-field/CHANGELOG.md
@@ -6,6 +6,11 @@ Unreleased
 ### Changed
 * Clear button added to search field for cross-browser compatibility
 
+3.12.0 - (May 1, 2019)
+------------------
+### Changed
+* Minor dependency version bump
+
 3.11.0 - (April 16, 2019)
 ------------------
 ### Changed